# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Venue {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  location?: string | null;
  timezone: string;
}

export interface Patron {
  id: string;
  venue_id: string;
  display_name: string;
  /** @nullable */
  email?: string | null;
  session_token: string;
  total_xp: number;
  loyalty_tier: string;
  created_at?: string;
  last_seen_at?: string;
}

export interface PatronInput {
  venue_id: string;
  display_name: string;
  /** @nullable */
  email?: string | null;
  session_token: string;
}

export interface Game {
  id: string;
  venue_id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  max_duration: number;
  is_active: boolean;
  sort_order: number;
}

export interface ScoreInput {
  game_id: string;
  patron_id: string;
  venue_id: string;
  score: number;
  /** @nullable */
  duration_secs?: number | null;
}

export interface Score {
  id: string;
  game_id: string;
  patron_id: string;
  venue_id: string;
  score: number;
  /** @nullable */
  duration_secs?: number | null;
  session_date: string;
  created_at?: string;
}

export interface LeaderboardEntry {
  game_id: string;
  game_name: string;
  game_slug: string;
  display_name: string;
  loyalty_tier: string;
  best_score: number;
  plays: number;
  rank: number;
}

export interface MenuItem {
  id: string;
  venue_id: string;
  category: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  is_available: boolean;
  is_featured: boolean;
  is_special: boolean;
  sort_order: number;
}

export interface Event {
  id: string;
  venue_id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  performer?: string | null;
  event_date: string;
  /** @nullable */
  start_time?: string | null;
  /** @nullable */
  ticket_url?: string | null;
  /** @nullable */
  ticket_price?: number | null;
  is_featured: boolean;
  is_active: boolean;
}

export type GetPatronParams = {
session_token: string;
};

export type GetLeaderboardParams = {
game_slug?: string;
};

export type GetLeaderboardAlltimeParams = {
game_slug?: string;
};