# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: string;
  venue_id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  performer?: string | null;
  event_date: string;
  /** @nullable */
  start_time?: string | null;
  /** @nullable */
  ticket_url?: string | null;
  /** @nullable */
  ticket_price?: number | null;
  is_featured: boolean;
  is_active: boolean;
}