# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Game {
  id: string;
  venue_id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  max_duration: number;
  is_active: boolean;
  sort_order: number;
}