# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  game_id: string;
  game_name: string;
  game_slug: string;
  display_name: string;
  loyalty_tier: string;
  best_score: number;
  plays: number;
  rank: number;
}