# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: string;
  venue_id: string;
  category: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  is_available: boolean;
  is_featured: boolean;
  is_special: boolean;
  sort_order: number;
}