# lib/api-zod/src/generated/types/patron.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Patron {
  id: string;
  venue_id: string;
  display_name: string;
  /** @nullable */
  email?: string | null;
  session_token: string;
  total_xp: number;
  loyalty_tier: string;
  created_at?: string;
  last_seen_at?: string;
}