# lib/api-zod/src/generated/types/patronInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PatronInput {
  venue_id: string;
  display_name: string;
  /** @nullable */
  email?: string | null;
  session_token: string;
}