# lib/api-zod/src/generated/types/score.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Score {
  id: string;
  game_id: string;
  patron_id: string;
  venue_id: string;
  score: number;
  /** @nullable */
  duration_secs?: number | null;
  session_date: string;
  created_at?: string;
}