# lib/api-zod/src/generated/types/scoreInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreInput {
  game_id: string;
  patron_id: string;
  venue_id: string;
  score: number;
  /** @nullable */
  duration_secs?: number | null;
}