# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pup Club API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Venue {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  location?: string | null;
  timezone: string;
}